% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{pforeach}
\alias{pforeach}
\alias{pforeach-package}
\title{pforeach}
\usage{
pforeach(..., .c, .combine = c, .parallel = TRUE, .debug = !.parallel,
  .cores, .seed = NULL, .init, .final = NULL, .inorder = TRUE,
  .multicombine = FALSE, .maxcombine = if (.multicombine) 100 else 2,
  .errorhandling = c("pass", "stop", "remove"), .packages = NULL,
  .export = NULL, .noexport = NULL, .verbose = FALSE)
}
\arguments{
\item{...}{input for foreach.}

\item{.c}{the relief of .combine parameter. Additionally .c=list means list combine that is deafault on foreach.}

\item{.combine}{function that is used to process the tasks results as they generated.}

\item{.parallel}{if TRUE, execute parallel processing.}

\item{.debug}{if TRUE, execute sequential processing(not parallel).}

\item{.cores}{number of cores for paralell processing.}

\item{.seed}{random number seed.}

\item{.packages}{character vector of packages that the tasks depend on.}

\item{.export}{character vector of variables to export.}
}
\description{
Easy way to parallel processing in R

Parallelized foreach
}

