% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{pforeach}
\alias{pforeach}
\alias{pforeach-package}
\title{pforeach}
\usage{
pforeach(..., .combine = c, .parallel = TRUE, .debug = !.parallel, .cores,
  .seed = NULL, .export, .packages)
}
\arguments{
\item{...}{input for foreach.}

\item{.combine}{function that is used to process the tasks results as they generated.}

\item{.parallel}{if TRUE, execute parallel processing.}

\item{.debug}{if TRUE, execute sequential processing(not parallel).}

\item{.cores}{number of cores for paralell processing.}

\item{.seed}{random number seed.}

\item{.export}{character vector of variables to export.}

\item{.packages}{character vector of packages that the tasks depend on.}
}
\description{
Easy use parallel processing

Parallelized foreach
}

